% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_new.R
\name{data_check}
\alias{data_check}
\title{Checking the user's data frame}
\usage{
data_check(data)
}
\arguments{
\item{data}{Any R object but ultimately it is a data frame.}
}
\value{
The function returns TRUE if the data frame indicated by the \code{data} parameter is suitable for the calculation of price indices and returns FALSE otherwise.
}
\description{
The function checks if the argument \code{data} points to a data frame which is suitable for further price index calculation. In particular, the function checks whether the indicated data frame contains the required columns and whether they are of the appropriate type (if not, the function returns FALSE and an appropriate comment).
}
\examples{
data_check(milk)
data_check(iris)
}
