% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data_preparing}
\alias{data_preparing}
\title{Preparing a data set for further data processing or price index calculations}
\usage{
data_preparing(
  data,
  prodID = TRUE,
  codeIN = FALSE,
  codeOUT = FALSE,
  retID = TRUE,
  description = TRUE,
  additional = c()
)
}
\arguments{
\item{data}{The user's data frame to be prepared. It must contain columns: \code{time} (as Date or character type, allowed formats are, eg.: `2020-03` or `2020-12-28`), \code{prices} and \code{quantities} (as numeric). Optionally, it may contain columns: \code{prodID}, \code{codeIN}, \code{codeOUT}, \code{retID} (as numeric or character), \code{description} (as character) and other columns specified by the \code{additional} parameter.}

\item{prodID}{A logical value indicating whether the \code{prodID} column is considered while data preparing (if it is set to TRUE, then records with missing values in this column are deleted). The \code{prodID} column includes unique product IDs used for product matching (as numeric or character). It is not obligatory to consider this column while data preparing but it is required while price index calculating (to obtain it, please see \code{\link{data_matching}}.}

\item{codeIN}{A logical value, e.g. if there are retailer (internal) product codes (as numeric or character) written in \code{codeIN} column and there is a need to use that column while data preparing, then that parameter should be set to TRUE. Otherwise it is set to be FALSE.}

\item{codeOUT}{A logical value, e.g. if there are retailer (internal) product codes (as numeric or character) written in \code{codeIN} column and there is a need to use that column while data preparing, then that parameter should be set to TRUE. Otherwise it is set to be FALSE.}

\item{retID}{A logical value, e.g. if there are outlet (retailer) codes (as numeric or character) written in \code{retID} column and there is a need to use that column while data preparing, then that parameter should be set to TRUE. Otherwise it is set to FALSE.}

\item{description}{A logical value, e.g. if there are product labels (as character) written in \code{description} column and there is a need to use that column while data preparing, then that parameter should be set to TRUE. Otherwise it is set to FALSE.}

\item{additional}{A vector of names of additional columns to be considered while data preparing (records with missing values are deleted).}
}
\value{
The resulting data frame is free from missing values, zero or negative prices and quantities. As a result, column \code{time} is set to be Date type (in format: `Year-Month-01`), columns \code{prices} and \code{quantities} are set to be numeric. If the `description` parameter is set to TRUE, then the column \code{description} is set to be character type.
}
\description{
This function returns a prepared data frame based on the user's data set. The resulting data frame is ready for further data processing (such as data selecting, matching or filtering) and it is also ready for price index calculations (if only it contains required columns).
}
\examples{
data_preparing(milk,prodID=FALSE)
data_preparing(dataMATCH, prodID=FALSE, codeIN=TRUE, codeOUT=TRUE, description=FALSE)
}
