% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{data_selecting}
\alias{data_selecting}
\title{Selecting products from the user's data set for further price index calculations}
\usage{
data_selecting(data, include = c(), must = c(), exclude = c())
}
\arguments{
\item{data}{The user's data frame. It must contain a column \code{description} (as character).}

\item{include}{A vector consisting of words and phrases. The function reduces the data set to one in which the \code{description} column contains any of these values.}

\item{must}{A vector consisting of words and phrases. The function reduces the data set to one in which the \code{description} column contains each of these values.}

\item{exclude}{A vector consisting of words and phrases. The function reduces the data set to one in which the \code{description} column does not contain any of these values.}
}
\value{
The function returns a subset of the user's data set obtained by selection based on keywords and phrases defined by parameters: \code{include}, \code{must} and \code{exclude}. Providing values of these parameters, please remember that the procedure distinguishes between uppercase and lowercase letters.
}
\description{
The function returns a subset of the user's data set obtained by selection based on keywords and phrases.
}
\examples{
data_selecting(milk, include=c("milk"), must=c("UHT"))
data_selecting(milk, must=c("milk"), exclude=c("paust"))
}
