% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataDOWNSIZED.R
\docType{data}
\name{dataDOWNSIZED}
\alias{dataDOWNSIZED}
\title{An artificial data set on sold coffee}
\format{
A data frame with 6 columns and 22 rows. The used variables are as follows:

\code{time} - Dates of transactions (Year-Month-Day)

\code{prices} - Prices of sold products [PLN]

\code{quantities} - Quantities of sold products [liters]

\code{codeIN} - Unique internal product codes (retaler product codes)

\code{codeOUT} - Unique external product codes (e.g. GTIN, EAN, SKU)

\code{description} Descriptions of sold coffee products
}
\usage{
dataDOWNSIZED
}
\description{
A collection of scanner data on the sale of coffee in the period from January 2024 to February 2024
}
\keyword{datasets}
