% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataU.R
\docType{data}
\name{dataU}
\alias{dataU}
\title{An artificial, small scanner data set}
\format{
A data frame with 5 columns and 6 rows. The used variables are as follows:

\code{time} - Dates of transactions (Year-Month-Day)

\code{prices} - Prices of sold products [PLN]

\code{quantities} - Quantities of sold products [item]

\code{prodID} - Unique product codes 

\code{description} Descriptions of sold products (data set contains 6 different product descriptions)
}
\usage{
dataU
}
\description{
A collection of artificial scanner data on 6 products sold in Dec, 2018. Product descriptions contain the information about their grammage and unit.
}
\keyword{datasets}
