% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_new.R
\name{hybrid}
\alias{hybrid}
\title{Calculating the bilateral hybrid price index}
\usage{
hybrid(data, start, end, base = start, interval = FALSE)
}
\arguments{
\item{data}{The user's data frame with information about sold products. It must contain columns: \code{time} (as Date in format: year-month-day,e.g. '2020-12-01'), \code{prices} (as positive numeric), \code{quantities}  (as positive numeric) and \code{prodID} (as numeric or character).}

\item{start}{The base period (as character) limited to the year and month, e.g. '2020-03'.}

\item{end}{The research period (as character) limited to the year and month, e.g. '2020-04'.}

\item{base}{The prior period used in the hybrid price index formula (as character) limited to the year and month, e.g. '2020-01'.}

\item{interval}{A logical value indicating whether the function is to compare the research period defined by \code{end} to the base period defined by \code{start} (then \code{interval} is set to FALSE) or all fixed base indices are to be calculated. In this latter case, all months from the time interval \code{<start,end>} are considered and \code{start} defines the base period (\code{interval} is set to TRUE).}
}
\value{
The function returns a value (or a vector of values) of the bilateral hybrid price index depending on the \code{interval} parameter. If the \code{interval} parameter is set to TRUE, the function returns a vector of price index values without dates. To get information about both price index values and corresponding dates, please see functions: \code{\link{price_index}}, \code{\link{price_indices}}, \code{\link{final_index}} or \code{\link{final_index2}}. The function does not take into account aggregating over outlets or product subgroups (to consider these types of aggregating, please use the \code{\link{final_index}} or \code{\link{final_index2}} function).
}
\description{
This function returns a value (or a vector of values) of the bilateral hybrid price index. The hybrid index was proposed by Bialek (2020) and it uses correlation coefficients between prices and quantities.
}
\examples{
\donttest{hybrid(milk, start="2019-12", end="2020-08", base="2018-12")}
\donttest{hybrid(milk, start="2019-12", end="2020-08", base="2018-12", interval=TRUE)}
}
\references{
{Bialek, J. (2020). \emph{Proposition of a Hybrid Price Index Formula for the Consumer Price Index Measurement}. Equilibrium. Quarterly Journal of Economics and Economic Policy, 15(4), 697-716.}
}
